refactor uplinks
