+++ packages/proxy/src/index.ts
@@ -1,2 +1,1 @@
 export * from './up-storage';
-export * from './uplink-util';
