--- packages/server/test/proxy/proxy.yaml
@@ -1,20 +1,0 @@
-storage: ./storage
-plugins: ./plugins
-auth:
-  htpasswd:
-    file: ./htpasswd
-security:
-  api:
-    jwt:
-      sign:
-        expiresIn: 10m
-        notBefore: 0
-packages:
-  '@*/*':
-    access: $authenticated
-    publish: $authenticated
-  '**':
-    access: $authenticated
-    publish: $authenticated
-logs:
-  - { type: stdout, format: pretty, level: warn }
