+++ packages/store/src/uplink-util.ts
@@ -1,6 +1,6 @@
 import { Versions, Config } from '@verdaccio/types';
 
-import { ProxyStorage, IProxy, ProxyList } from './up-storage';
+import { ProxyStorage, IProxy, ProxyList } from '@verdaccio/proxy';
 
 /**
  * Set up the Up Storage for each link.
@@ -27,7 +27,6 @@ export function updateVersionsHiddenUpLink(versions: Versions, upLink: IProxy): 
       const version = versions[i];
 
       // holds a "hidden" value to be used by the package storage.
-      // $FlowFixMe
       version[Symbol.for('__verdaccio_uplink')] = upLink.upname;
     }
   }
